refactor(management): extract route operations
